all images endpoint done

+++ spadeAPI/controllers/tenantController.js
@@ -433,68 +433,51 @@ exports.tenantDelete = async (req, res) => {
     const { tenantID } = req.body
     const tenantResult = await queryRunner(selectQuery("tenants", "id"), [tenantID]);
     // console.log(tenantResult[0][0])
-    const propertyUnitID = tenantResult[0][0].propertyUnitID;
-    const tenantDeleteResult = await queryRunner(deleteQuery("tenants", "id"), [tenantID]);
-    // console.log(tenantDeleteResult[0])
-    if (tenantDeleteResult[0].affectedRows > 0) {
+    if (tenantResult[0].length > 0) {
+      const propertyUnitID = tenantResult[0][0].propertyUnitID;
+      const tenantDeleteResult = await queryRunner(deleteQuery("tenants", "id"), [tenantID]);
+      // console.log(tenantDeleteResult[0])
+      if (tenantDeleteResult[0].affectedRows > 0) {
 
-      const tenantCheckResult = await queryRunner(selectQuery("tenantattachfiles", "tenantID"), [tenantID]);
-      // console.log(tenantCheckResult)
-      if (tenantCheckResult[0].length > 0) {
-        tenantimages = tenantCheckResult[0].map((image) => image.fileName);
-        // delete folder images
-        imageToDelete(tenantimages);
-        const tenantFileDeleteresult = await queryRunner(deleteQuery("tenantattachfiles", "tenantID"), [tenantID]);
-        // console.log(tenantFileDeleteresult
-      }
-      const tenantAdditionalEmailresult = await queryRunner(deleteQuery("tenantalternateemail", "tenantID"), [tenantID]);
-      if (tenantAdditionalEmailresult[0].affectedRows > 0) {
-        const tenantAdditionalPhoneResult = await queryRunner(deleteQuery("tenantalternatephone", "tenantID"), [tenantID]);
-        if (tenantAdditionalPhoneResult[0].affectedRows > 0) {
-          const tenantIncreaseRentResult = await queryRunner(deleteQuery("tenantincreaserent", "tenantID"), [tenantID]);
-
-
-          if (tenantIncreaseRentResult[0].affectedRows > 0) {
-            //dddddd
-            const status = "Vacant";
-            const propertyUnitsResult = await queryRunner(updateUnitsTenant, [status, propertyUnitID]);
-
-
-            if (propertyUnitsResult[0].affectedRows > 0) {
-
-              res.status(200).json({
-                message: " tenant deleted successfully"
-              })
-            } else {
-              // " tenant Error occur in increase rent "
-              res.status(200).json({
-                message: " tenant Error occur in updating property units "
-              })
-            }
-          } else {
-            // " tenant Error occur in increase rent "
-            res.status(200).json({
-              message: " tenant Error occur in delete increase rent "
-            })
-          }
-          // }
-
-        } else {
-          res.status(200).json({
-            message: " tenant Error occur in delete alternate phone "
-          })
+        const tenantCheckResult = await queryRunner(selectQuery("tenantattachfiles", "tenantID"), [tenantID]);
+        if (tenantCheckResult[0].length > 0) {
+          const tenantimages = tenantCheckResult[0].map((image) => image.fileName);
+          // delete folder images
+          imageToDelete(tenantimages);
+          const tenantFileDeleteresult = await queryRunner(deleteQuery("tenantattachfiles", "tenantID"), [tenantID]);
         }
 
+        const tenantAdditionalEmailCheckResult = await queryRunner(selectQuery("tenantalternateemail", "tenantID"), [tenantID]);
+        if (tenantAdditionalEmailCheckResult[0].length > 0) {
+        const tenantAdditionalEmailresult = await queryRunner(deleteQuery("tenantalternateemail", "tenantID"), [tenantID]);
+        } 
 
-      } else {
+
+        const tenantAdditionalPhoneCheckResult = await queryRunner(selectQuery("tenantalternatephone", "tenantID"), [tenantID]);
+        if (tenantAdditionalPhoneCheckResult[0].length > 0) {
+          const tenantAdditionalPhoneResult = await queryRunner(deleteQuery("tenantalternatephone", "tenantID"), [tenantID]);
+        }
+
+        const tenantIncreaseRentCheckResult = await queryRunner(selectQuery("tenantincreaserent", "tenantID"), [tenantID]);
+        if (tenantIncreaseRentCheckResult[0].length > 0) {
+          const tenantIncreaseRentResult = await queryRunner(deleteQuery("tenantincreaserent", "tenantID"), [tenantID]);
+        }
+
+          const status = "Vacant";
+          const propertyUnitsResult = await queryRunner(updateUnitsTenant, [status, propertyUnitID]);
+        
+        
         res.status(200).json({
-          message: " tenant Error occur in delete alternate email "
+          message: " tenant deleted successfully"
+        })
+      } else { // tenantCheckResult
+        res.status(200).json({
+          message: "Error occur in delete tenant "
         })
       }
-
     } else { // tenantCheckResult
       res.status(200).json({
-        message: "No tenant file data found "
+        message: "No tenant found "
       })
     }
